ai-service/requirements.txt
fastapi
uvicorn[standard]
pymupdf
google-genai
Pillow
python-docx
pydantic
python-multipart
python-dotenv
